backend/requirements.txt
# --- Web & API ---
fastapi==0.135.1
uvicorn==0.41.0
python-dotenv==1.2.2
python-multipart

# --- Task Queue ---
celery==5.3.6
redis==5.0.1

# --- Database & RAG ---
psycopg2-binary
pgvector
supabase

# --- Storage ---
boto3
google-cloud-storage

# --- AI & Models ---
google-genai
deepgram-sdk==3.2.7

# --- Video & Audio Processing ---
librosa
soundfile
numpy
mediapipe
opencv-python-headless

scenedetect[opencv]==0.6.5.2

# --- YouTube Tools ---
yt-dlp>=2026.3.3

# --- Director Module ---
sentry-sdk[fastapi]
langfuse>=2.0.0
posthog